Add support for octavia
If the cloud uses octavia and not neutron lbaas, you need a flag in
cloud-config.

The k8s docs indicate that use-octavia will fall back to looking for
neutron-lbaas if it can't find octavia - maybe it would not be a
terrible idea to just default it to true?

Add security-group rules for octavia

The port-range of the load balancers need to be opened so that octavia
load balancers can talk to the service ips on the nodes.