Only add ssh keys if they do not exist
The current ssh known_hosts code adds the host keys unconditionally
every time the ansible is run. This can result in quite a lot of ssh
keys repeatedly added to known_hosts if, for instance, the ansible
is run in a cron job.

Instead of >>, use lineinfile and the pipe lookup filter. This is
already done for nodes.